backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
diffusers==0.30.3
transformers==4.45.2
accelerate==1.0.1
pillow==10.4.0
requests==2.32.3
python-multipart==0.0.9
